refactor: Move the custom portal typescript language stuff to the public folder so that it will be placed in the monaco chunk
