+++ src/aux-server/aux-web/shared/public/monaco-editor/custom-portal-typescript/custom-portal-typescript.ts
