+++ src/aux-server/aux-web/shared/public/monaco-editor/languages.contribution.ts
