+++ src/aux-server/aux-web/shared/public/monaco-editor/typescript/index.ts
